ci: Disable manual run on limited test workflow (#33534)
The step to get the PR number when run manually has error in its code.
The `res=` line doesn't work, in that it'll mangle the JSON output from
the `curl` command, before setting it into the variable. It should've
been

```sh
res="$(curl ...)"
```

Considering we don't use this functionality today, and it doesn't work
either, we're removing it.

This workflow is only used with `repository_dispatch`, and we make that
explicit.