chore: removed constraint on order of execution for consolidated apis (#34140)
## Description
- In the last implementation of auto-commit we were comparing dslVersion
of page dsl object with rts's latest version as a criteria to trigger
auto-commit. This check was implemented in the page fetch section of
`consolidated api call (page load)` itself. Since page dsls are migrated
in the consolidated api, hence we required this check to be executed
before the pages of application were migrated. Now that is not the case
anymore. hence order of execution for home page load is not a constraint
anymore. A removal is in order.

- Added a feature flag annotation with
`release_git_autocommit_feature_enabled` to stop feature flag leaks on
publisheAutocommitEvent method call.

Fixes #`Issue Number`  
_or_  
Fixes `Issue URL`
> [!WARNING]  
> _If no issue exists, please create an issue first, and check with the
maintainers if the issue is valid._

## Automation

/ok-to-test tags="@tag.All"

### :mag: Cypress test results
<!-- This is an auto-generated comment: Cypress test results  -->
> [!IMPORTANT]
> 🟣 🟣 🟣 Your tests are running.
> Tests running at:
<https://github.com/appsmithorg/appsmith/actions/runs/9448324929>
> Commit: 7f7f88fcd8a761aa9deacdfcb1a93bdc6d5f9900
> Workflow: `PR Automation test suite`
> Tags: ``

<!-- end of auto-generated comment: Cypress test results  -->




## Communication
Should the DevRel and Marketing teams inform users about this change?
- [ ] Yes
- [ ] No